Разбили код на модули и добавили пустые файлы для будущих моделей
